Moved enum to another file
