+++ packages/api/src/routers/auth/types.ts
@@ -1,0 +1,4 @@
+export enum AuthStatus {
+  SignedIn = 'SignedIn',
+  PhoneAlreadyUsed = 'PhoneAlreadyUsed',
+}
